Fixed resource typos for blueberries and lingonberries
Noticed there was just typos in the berries cat for blueberries and lingonberries, Berry and Forageable both have stacksizes of 4. 

stinging nettle was added but its not really needed so removed that as well to avoid confusion as its added in a few other places already plus its got a customstacksize.